+++ src/frequenz/channels/util/_file_watcher.py
@@ -10,19 +10,18 @@ from typing import List, Optional, Set, Union
 from watchfiles import Change, awatch
 from watchfiles.main import FileChange
 
-from frequenz.channels.base_classes import ChannelClosedError, Receiver
-
-
-class EventType(Enum):
-    """Available types of changes to watch for."""
-
-    CREATE = Change.added
-    MODIFY = Change.modified
-    DELETE = Change.deleted
+from .._base_classes import ChannelClosedError, Receiver
 
 
 class FileWatcher(Receiver[pathlib.Path]):
     """A channel receiver that watches for file events."""
+
+    class EventType(Enum):
+        """Available types of changes to watch for."""
+
+        CREATE = Change.added
+        MODIFY = Change.modified
+        DELETE = Change.deleted
 
     def __init__(
         self,
@@ -37,7 +36,7 @@ class FileWatcher(Receiver[pathlib.Path]):
                 all event types.
         """
         if event_types is None:
-            event_types = {EventType.CREATE, EventType.MODIFY, EventType.DELETE}
+            event_types = set(FileWatcher.EventType)  # all types
 
         self.event_types = event_types
         self._stop_event = asyncio.Event()